// frontend/src/types/types.ts
export interface PerfumeResponse {
    id: number;
    perfumeTitle: string;
    perfumer: string;
    price: number;
    perfumeRating: number;
    filename: string;
    reviewsCount: number;
    volume: string;
}
// ...
export interface OrderResponse {
    id: number;
    totalPrice: number;
    date: string;
    firstName: string;
    lastName: string;
    city: string;
    address: string;
    email: string;
    phoneNumber: string;
    postIndex: number;
    url: string;
}

export interface OrderItemResponse {
    id: number;
    amount: number;
    quantity: number;
    perfume: PerfumeResponse;
}

export interface OrderError {
    emailError: string;
    firstNameError: string;
    lastNameError: string;
    cityError: string;
    addressError: string;
    postIndexError: string;
    phoneNumberError: string;
}
// ...
export enum LoadingStatus {
    LOADED = "LOADED",
    LOADING = "LOADING",
    ERROR = "ERROR",
    NEVER = "NEVER",
    SUCCESS = "SUCCESS"
}